Update invokable.hpp
New implementation did not support argument forwarding on invoke(). Previously the forwarding was *attempted* on callback constructor, but that doesn't actually forward arguments at runtime when you call invoke or invokable<A...>.invoke() where the rvalues decay into lvalues when passed from invokable<A..>.invoke() to callback<A...>.invoke().

Wrapping the function call in a lambda prior was necessary for variadic template arguments, but not anymore with the use of std::function, now we just call std::move() to move the function call reference to callback<A...> internal function call.