Update invokable.hpp
Hashes don't evaluate properly due to lambda usage, which is the primary usage here. So we do not check for duplicates on hook() or on std::vector insert due to this. However we can still use hashes to call unhook() or std::vector erase_if for matching hashes of existing callbacks.